requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==41.0.3
Django==3.2
django-admin-rangefilter==0.9.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-environ==0.10.0
djangorestframework==3.14.0
more-itertools==10.1.0
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
